Move fileStep to directory + refector files
