+++ src/components/quote/QuoteContent.tsx
@@ -2,7 +2,7 @@ import { useContext, useEffect, useRef } from "react"
 import { SliderStep } from "./steps/SliderStep"
 import { QuoteContext } from "./_context/QuoteContext"
 import { CatalogStep } from "./steps/CatalogStep"
-import { FilesStep } from "./steps/FilesStep"
+import { FilesStep } from "./steps/FilesStep/FilesStep"
 import { ReactLenis, useLenis } from "@studio-freight/react-lenis"
 import { useWindowSize } from "@uidotdev/usehooks"
 import { Recap } from "./steps/recap/Recap"
